Fix or silence (msvc) compiler warnings about narrowing conversions, unused formal parameters, constant conditional expressions, name hiding, dll-interfaces and trivias

+++ test/write_failure_test.cpp
@@ -30,7 +30,7 @@ struct limit_device {
             }
             n /= 2;
         }
-        written += n;
+        written += static_cast<int>(n);
         return n;
     }
 };
